pyproject.toml
[tool.poetry]
name = "issue-tracker-bot"
version = "0.1.0"
description = ""
authors = ["Oleksandr Marusyk <redcyb@gmail.com>"]
readme = "README.md"
packages = [{include = "issue_tracker_bot"}]

[tool.poetry.dependencies]
python = "~3.7"
flask = "^2.2.5"
gunicorn = "^21.2.0"
requests = "^2.31.0"
boto3 = "^1.28.35"
pydantic = "^2.3.0"
google-api-python-client = "^2.97.0"
google-auth-httplib2 = "^0.1.0"
google-cloud-api-keys = "^0.5.3"
google-auth-oauthlib = "^1.0.0"
python-dotenv = "^0.21.1"
python-telegram-bot = {extras = ["all"], version = "^20.3"}
fastapi = "^0.103.1"
uvicorn = {version = "0.22.0", extras = ["standard"]}
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
sqlalchemy-utils = "^0.41.1"

[tool.poetry.group.dev]

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pytest = "^7.4.0"
pytest-env = "^1.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
